clean up code remove unnecessary methods move AbstractGate class to abstract.py
